ml/pyproject.toml
[project]
name = "llm-tycoon-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.93.0",
    "fastapi[standard]>=0.135.3",
    "numpy>=2.4.4",
    "pydantic-settings>=2.13.1",
    "torch>=2.11.0",
]
